Extract console report
Closing https://github.com/fastruby/skunk/pull/103
