+++ test/samples/console_output.txt
@@ -1,12 +1,1 @@
-+---------------------------------------+----------------+------------------+--------------+--------------+--------------+
-| file                                  | skunk_score    | churn_times_cost | churn        | cost         | coverage     |
-+---------------------------------------+----------------+------------------+--------------+--------------+--------------+
-| samples/rubycritic/analysed_module.rb | 0.59           | 0.59             | 1            | 0.59         | 100.0        |
-+---------------------------------------+----------------+------------------+--------------+--------------+--------------+
-
-SkunkScore Total: 0.59
-Modules Analysed: 1
-SkunkScore Average: 0.59
-Worst SkunkScore: 0.59 (samples/rubycritic/analysed_module.rb)
-
-Generated with Skunk v0.5.2
+Score: 96.32
